+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - `ShlexParser` no-longer treats `'` as a quote.
+- Command objects can now be passed directly to `SlashCommand.__init__` and `MessageCommand.__init__`.
 
 ## [2.3.1a1] - 2022-01-27
 ### Added